requirements.txt
Flask==2.3.3
twilio==8.3.0
gunicorn==23.0.0
python-dotenv==1.0.0